+++ macfiles-native/src/main/kotlin/garden/ephemeral/macfiles/native/NativeCommon.kt
@@ -1,9 +1,9 @@
 package garden.ephemeral.macfiles.native
 
-import com.sun.jna.Platform
-import garden.ephemeral.macfiles.native.internal.SystemB
-import java.io.File
-import java.text.Normalizer
+import com.sun.jna.*
+import garden.ephemeral.macfiles.native.internal.*
+import java.io.*
+import java.text.*
 
 internal fun getVolumePath(file: File): File {
     // Find the filesystem
@@ -17,5 +17,12 @@ internal fun getVolumePath(file: File): File {
 internal fun requireMacOS() {
     if (!Platform.isMac()) {
         throw UnsupportedOperationException("Not implemented on this platform (requires native macOS support)")
+    }
+}
+
+@Throws(NoSuchFileException::class)
+internal fun checkFileExists(file: File) {
+    if (!file.exists()) {
+        throw NoSuchFileException(file)
     }
 }